flask-jwt to flask-jwt-extend and Added JWT_Claims & jwt identity & jwt_optional

+++ resources/item.py
@@ -1,5 +1,5 @@
 from flask_restful import Resource, reqparse
-from flask_jwt import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt_claims, jwt_optional, get_jwt_identity
 from models.item import ItemModel
 
 
@@ -16,7 +16,7 @@ class Item(Resource):
                         help="Every item needs a store_id!"
                         )
 
-    @jwt_required()
+    @jwt_required
     def get(self, name):
         item = ItemModel.find_by_name(name)
         if item:
@@ -37,15 +37,19 @@ class Item(Resource):
 
         return item.json(), 200
 
-    @jwt_required()
+    @jwt_required
     def delete(self, name):
+        claim = get_jwt_claims()
+        if not claim['is_admin']:
+            return {'message': 'Admin privilage required.'}, 401
+
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
 
         return {'message': 'Item delete'}
 
-    @jwt_required()
+    @jwt_required
     def put(self, name):
         data = Item.parser.parse_args()
         item = ItemModel.find_by_name(name)
@@ -60,6 +64,13 @@ class Item(Resource):
 
 
 class ItemList(Resource):
+    @jwt_optional
     def get(self):
-        # return {'items': list(map(lambda item: item.json(), ItemModel.query.all()))}
-        return {'items': [item.json() for item in ItemModel.find_all()]}
+        user_id = get_jwt_identity()
+        items = [item.json() for item in ItemModel.find_all()]
+        if user_id:
+            return {'items': items}, 200
+        return {
+            'items': [item['name'] for item in items], 
+            'message': 'More data available if you login'
+        }, 200
